+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to semantic versioning.
 ## 0.7.0 (_Unreleased_)
 
 - Add `aperture` function to the `Array` module.
+- Add `construct` function to the `Function` module.
 - Provide the input value to the fallback function of `guard`.
 
 ## 0.6.0 (2020-11-20)